pets refactored to standard step in form flow
