+++ src/components/pages/IntakePacketContent/ByGuests/Pets.js
@@ -4,13 +4,7 @@ This component contains:
   -Pet Information(input(s))
  
 */
-
-import React, { useEffect, useState } from 'react';
-import axios from 'axios';
-import { useHistory, Route } from 'react-router-dom';
-import { useDispatch, useSelector } from 'react-redux';
-import { axiosWithAuth } from '../../../../api/axiosWithAuth';
-import { getDocuSignUrl } from '../../../../state/actions/index';
+import React from 'react';
 import IntakeButton from '../IntakeButtons';
 
 //Ant Design imports (https://ant.design/components/overview/)
@@ -24,84 +18,18 @@ const Pets = ({
   steps,
   step,
 }) => {
-  //docusign
-  const signerInfo = useSelector(state => state.SIGNER_INFORMATION);
-  let envelopeArgs = {
-    signer1Email: signerInfo.email,
-    signer1Name: signerInfo.first_name + ' ' + signerInfo.last_name,
-    signer1Id: signerInfo.id,
-  };
-  const [loadDocuSign, setLoadDocusign] = useState(false);
-  const { familyInfo, familyMember } = formData;
-  const { previous } = navigation;
-  const history = useHistory();
-  const dispatch = useDispatch();
-  useEffect(() => {
-    if (loadDocuSign) {
-      history.push('/redirect');
-    }
-  }, [history, loadDocuSign]);
-
-  function callDocusign() {
-    // Saves family information so it does not get deleted after redirecting
-    axiosWithAuth()
-      .post(`/families`, familyInfo)
-      .then(res => {
-        const familyId = res.data.families.id;
-        Object.keys(formData.familyMember).map(mem => {
-          familyMember[mem]['family_id'] = familyId;
-          axiosWithAuth()
-            .post('/members', familyMember[mem])
-            .then(res => console.log('Members added', res.data))
-            .catch(err => {
-              console.log('MemberError', err.response);
-            });
-        });
-        // This axios post request calls the eg001.createController function in the backend
-        axios
-          .post('http://localhost:8000/callDS', envelopeArgs)
-          .then(res => {
-            setLoadDocusign(!loadDocuSign);
-            dispatch(getDocuSignUrl(res.data));
-          })
-          .catch(err => console.log('DocuSign error', err));
-      })
-      .catch(err => {
-        console.log('FamiliesError', err);
-      });
-  }
-  //docusign
-  const redirectToDocusign = async () => {
-    try {
-      const res = await axios.post(
-        'http://localhost:8000/callDS',
-        envelopeArgs
-      );
-      setLoadDocusign(!loadDocuSign);
-      dispatch(getDocuSignUrl(res.data));
-    } catch (error) {
-      console.log('Error in load docusign', error);
-    }
-  };
-
   //Progress bar
   const pageNumber = steps.findIndex(item => item === step);
   const pages = steps.length;
   const percent = ((pageNumber + 1) / pages) * 100;
-
+  const { familyInfo } = formData;
   return (
     <div style={tempFormStyle}>
       {/*Progress bar*/}
       <Progress percent={percent} status="active" showInfo={false} />
 
       <Card title="Pets" bordered={false}>
-        <div
-          style={{
-            display: 'flex',
-            justifyContent: 'space-between',
-            marginBottom: '30px',
-          }}
-        >
+        <div>
           <IntakeButton navigation={navigation} />
         </div>
